requirements.txt
Flask==3.1.1
influxdb==5.3.2
MarkupSafe==3.0.2
pandas==2.3.0
python-dotenv==1.1.1
Requests==2.32.4
Flask-SQLAlchemy==3.1.1
APScheduler==3.11.0
openpyxl==3.1.5
psutil==7.0.0
